requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1